Totally refactored sets #61
Taking an opportunity to refactor sets into models and collections that
only deal with sets.
