+++ app/exercises/models/exercise.js
@@ -4,22 +4,6 @@ import Backbone from 'backbone';
 import moment from 'moment';
 
 
-export const SetModel = Backbone.Model.extend({
-  defaults: {
-    'workout_date': ''
-  },
-
-  formatDate: function() {
-    return moment(this.get('workout_date')).calendar(null, {
-      sameDay: '[Today]',
-      lastDay: '[Yesterday]',
-      lastWeek: '[Last] dddd',
-      sameElse: 'DD/MM/YYYY'
-    });
-  }
-});
-
-
 export const ExerciseModel = Backbone.Model.extend({
   defaults: () => ({
     exercise_name: '',
